+++ crates/protocol/src/user/wasm.rs
@@ -28,14 +28,16 @@ use crate::KeyParams;
 pub async fn run_dkg_protocol(
     validators_info_js: ValidatorInfoArray,
     user_signing_secret_key: Vec<u8>,
+    block_number: u32,
 ) -> Result<KeyShare, Error> {
     let validators_info = parse_validator_info(validators_info_js)?;
 
     let user_signing_keypair = sr25519_keypair_from_secret_key(user_signing_secret_key)?;
 
-    let key_share = user_participates_in_dkg_protocol(validators_info, &user_signing_keypair)
-        .await
-        .map_err(|err| Error::new(&format!("{}", err)))?;
+    let key_share =
+        user_participates_in_dkg_protocol(validators_info, &user_signing_keypair, block_number)
+            .await
+            .map_err(|err| Error::new(&format!("{}", err)))?;
 
     Ok(KeyShare(key_share))
 }