+++ crates/threshold-signature-server/src/signing_client/api.rs
@@ -39,14 +39,20 @@ use entropy_shared::{KeyVisibility, OcwMessageProactiveRefresh, SETUP_TIMEOUT_SE
 use parity_scale_codec::Decode;
 use sp_core::Pair;
 use subxt::{
-    backend::legacy::LegacyRpcMethods, ext::sp_core::sr25519, tx::PairSigner,
-    utils::AccountId32 as SubxtAccountId32, OnlineClient,
+    backend::legacy::LegacyRpcMethods,
+    ext::sp_core::sr25519,
+    tx::PairSigner,
+    utils::{AccountId32 as SubxtAccountId32, Static},
+    OnlineClient,
 };
 use synedrion::KeyShare;
 use tokio::time::timeout;
 
 use crate::{
-    chain_api::{entropy, get_api, get_rpc, EntropyConfig},
+    chain_api::{
+        entropy::{self, runtime_types::pallet_staking_extension::pallet::RefreshInfo},
+        get_api, get_rpc, EntropyConfig,
+    },
     helpers::{
         launch::LATEST_BLOCK_NUMBER_PROACTIVE_REFRESH,
         substrate::{
@@ -125,6 +131,7 @@ pub async fn proactive_refresh(
                     &app_state.listener_state,
                     encoded_key,
                     deserialized_old_key,
+                    ocw_data.block_number,
                 )
                 .await?;
                 let serialized_key_share = key_serialize(&new_key_share)
@@ -183,11 +190,12 @@ pub async fn do_proactive_refresh(
     state: &ListenerState,
     verifying_key: Vec<u8>,
     old_key: KeyShare<KeyParams>,
+    block_number: u32,
 ) -> Result<KeyShare<KeyParams>, ProtocolErr> {
     tracing::debug!("Preparing to perform proactive refresh");
     tracing::debug!("Signing with {:?}", &signer.signer().public());
 
-    let session_id = SessionId::ProactiveRefresh(verifying_key);
+    let session_id = SessionId::ProactiveRefresh { verifying_key, block_number };
     let account_id = SubxtAccountId32(signer.signer().public().0);
     let mut converted_validator_info = vec![];
     let mut tss_accounts = vec![];
@@ -272,7 +280,11 @@ pub async fn validate_proactive_refresh(
         .await?
         .ok_or_else(|| ProtocolErr::ChainFetch("Error getting Proactive Refresh data"))?;
     let mut hasher_chain_data = Blake2s256::new();
-    hasher_chain_data.update(ocw_data.encode());
+    let ocw_data_refresh_info = RefreshInfo {
+        proactive_refresh_keys: ocw_data.proactive_refresh_keys.clone(),
+        validators_info: ocw_data.validators_info.clone().into_iter().map(Static).collect(),
+    };
+    hasher_chain_data.update(ocw_data_refresh_info.encode());
     let chain_data_hash = hasher_chain_data.finalize();
     let mut hasher_verifying_data = Blake2s256::new();
     hasher_verifying_data.update(proactive_info.encode());